requirements/_base.txt
Django>1.8,<1.10.0
django-compressor==2.0
celery>=3.1.18,<3.2
django-celery==3.1.17
Unidecode
lxml
Pillow==2.9.0
redis==2.9.1
pika
unipath
django-braces
python-ooxml
django-rest-swagger==2.1.0
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.0.0
django-logtailer==1.0.1
Ebooklib==0.16
